0.6.0-Pre-Alpha ├─• Integracion completa del modulo de ofertas con el frontend: disponibilidad, nombre real del trabajador y ubicacion de referencia. ├─• Endpoints dedicados para activar y desactivar disponibilidad por PATCH independiente, sin requerir edicion completa. ├─• Permisos ampliados para TRABAJADOR y PREMIUM en edicion y cambio de disponibilidad, eliminando bloqueos 403. ├─• Ofertas nuevas parten con disponible = false; el trabajador las activa manualmente. ├─• OfertaServicioResponseDTO enriquecido con nombreTrabajador, usernameTrabajador, ubicacionReferencia, latitud, longitud y rangoDisponibilidadM. ├─• UsuarioUbicacionDTO corregido con @JsonProperty para deserializar rango_disponibilidad_m y apellidos desde usuarios_api. ├─• Ubicacion publica de oferta reducida a comuna y region; calle y numero nunca se exponen en la respuesta. ├─• Catalogo de estados expuesto via EstadoController y EstadoCatalogoService para consumo del frontend. └─• Seed de datos refactorizado con archivos externos en resources/seed/ e inicializacion idempotente ampliada.

+++ src/main/java/com/contrabajo/servicios_api/repository/EstadoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EstadoRepository extends JpaRepository<Estado, Integer> {
+public interface EstadoRepository extends JpaRepository<Estado, Short> {
     // Útiles para buscar el estado exacto sin saber su ID
     Optional<Estado> findByCodigo(String codigo);
     Optional<Estado> findByNombre(String nombre);
